git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@3291 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ inc/container.h
@@ -12,6 +12,6 @@ struct cobj_ref {
 	((struct cobj_ref) { (container), (object) } )
 
 // Should be such that CT_QUOTA_INF + CT_QUOTA_INF > CT_QUOTA_INF.
-#define CT_QUOTA_INF	(1UL << 62)
+#define CT_QUOTA_INF	(((uint64_t) 1) << 62)
 
 #endif
